1.修改项目管理的展示形式，以项目名称分组的形式展示。 2.处理服务端导入的输入项目_id的问题。 3.处理项目金额费用的类型问题。 4.修改项目的重新启用方式。 5.修改前端项目在网页内标签的名称。 6.注释结算模块。 7.重新调整权限对应菜单关系。

+++ pm-resource-server/src/graphql/def.ts
@@ -51,14 +51,14 @@ type Project {
   projStatus: String!
   contStatus: String!
   acceStatus: String!
-  contAmount: Int
-  recoAmount: Int
-  projBudget: Int
-  budgetFee: Int
-  budgetCost: Int
-  actualFee: Int
-  actualCost: Int
-  taxAmount: Int
+  contAmount: Float
+  recoAmount: Float
+  projBudget: Float
+  budgetFee: Float
+  budgetCost: Float
+  actualFee: Float
+  actualCost: Float
+  taxAmount: Float
   description: String
   createDate: String!
   updateTime: String
@@ -306,14 +306,14 @@ input ProjectInput {
   projStatus: String!
   contStatus: String!
   acceStatus: String!
-  contAmount: Int
-  recoAmount: Int
-  projBudget: Int
-  budgetFee: Int
-  budgetCost: Int
-  actualFee: Int
-  actualCost: Int
-  taxAmount: Int
+  contAmount: Float
+  recoAmount: Float
+  projBudget: Float
+  budgetFee: Float
+  budgetCost: Float
+  actualFee: Float
+  actualCost: Float
+  taxAmount: Float
   description: String
   participants: [String!]
   contacts: [ContactInput!]
